Update history in redux state after post wellbeing
The post to wellbeings returns the created wellbeing post.

The wellbeings history is now in redux state rather than component
state.
For this reason this can be used to update history in the state,
so that it will be shown as soon as the app navigates back to
the home page.